Introduce common file, rename non-common files, remove stub
