+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/SqlCommand.netcore.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
@@ -12,29 +13,21 @@ using System.Data.SqlTypes;
 using System.Diagnostics;
 using System.IO;
 using System.Runtime.CompilerServices;
-using System.Runtime.ConstrainedExecution;
-using System.Security.Permissions;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Xml;
-using System.Buffers;
 using Microsoft.Data.Common;
 using Microsoft.Data.Sql;
+using Microsoft.Data.SqlClient.Diagnostics;
 using Microsoft.Data.SqlClient.Server;
-using System.Transactions;
-using System.Collections.Concurrent;
 
 // NOTE: The current Microsoft.VSDesigner editor attributes are implemented for System.Data.SqlClient, and are not publicly available.
 // New attributes that are designed to work with Microsoft.Data.SqlClient and are publicly documented should be included in future.
 namespace Microsoft.Data.SqlClient
 {
-    /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/SqlCommand/*'/>
-    [DefaultEvent("RecordsAffected")]
-    [ToolboxItem(true)]
-    [DesignerCategory("")]
     // TODO: Add designer attribute when Microsoft.VSDesigner.Data.VS.SqlCommandDesigner uses Microsoft.Data.SqlClient
-    public sealed class SqlCommand : DbCommand, ICloneable
+    public sealed partial class SqlCommand : DbCommand, ICloneable
     {
         private static int _objectTypeCount; // EventSource Counter
         private const int MaxRPCNameLength = 1046;
@@ -162,6 +155,10 @@ namespace Microsoft.Data.SqlClient
         /// </summary>
         private static bool _forceRetryableEnclaveQueryExecutionExceptionDuringGenerateEnclavePackage = false;
 #endif
+
+        private static readonly SqlDiagnosticListener s_diagnosticListener = new SqlDiagnosticListener();
+        private bool _parentOperationStarted = false;
+
         internal static readonly Action<object> s_cancelIgnoreFailure = CancelIgnoreFailureCallback;
 
         // Prepare
@@ -176,8 +173,6 @@ namespace Microsoft.Data.SqlClient
             PREPARED,           // execute prepared commands, all server versions   (results in sp_exec call)
         }
 
-        // devnotes
-        //
         // _hiddenPrepare
         // On 8.0 and above the Prepared state cannot be left. Once a command is prepared it will always be prepared.
         // A change in parameters, commandtext etc (IsDirty) automatically causes a hidden prepare
@@ -327,7 +322,7 @@ namespace Microsoft.Data.SqlClient
             internal void SetActiveConnectionAndResult(TaskCompletionSource<object> completion, string endMethod, SqlConnection activeConnection)
             {
                 Debug.Assert(activeConnection != null, "Unexpected null connection argument on SetActiveConnectionAndResult!");
-                TdsParser parser = activeConnection.Parser;
+                TdsParser parser = activeConnection?.Parser;
                 SqlClientEventSource.Log.TryTraceEvent("SqlCommand.SetActiveConnectionAndResult | API | ObjectId {0}, Client Connection Id {1}, MARS={2}", activeConnection?.ObjectID, activeConnection?.ClientConnectionId, parser?.MARSOn);
                 if ((parser == null) || (parser.State == TdsParserState.Closed) || (parser.State == TdsParserState.Broken))
                 {
@@ -336,7 +331,7 @@ namespace Microsoft.Data.SqlClient
 
                 _cachedAsyncCloseCount = activeConnection.CloseCount;
                 _cachedAsyncResult = completion;
-                if (activeConnection != null && !parser.MARSOn)
+                if (!parser.MARSOn)
                 {
                     if (activeConnection.AsyncCommandInProgress)
                         throw SQL.MARSUnsupportedOnConnection();
@@ -377,7 +372,6 @@ namespace Microsoft.Data.SqlClient
         private int _rowsAffectedBySpDescribeParameterEncryption = -1;
 
         private SqlNotificationRequest _notification;
-        private bool _notificationAutoEnlist = true;            // Notifications auto enlistment is turned on by default
 
         // transaction support
         private SqlTransaction _transaction;
@@ -515,7 +509,6 @@ namespace Microsoft.Data.SqlClient
                             // cleanup
                             Unprepare();
                         }
-                        // @TODO: CER Exception Handling was removed here (see GH#3581)
                         catch (Exception)
                         {
                             // we do not really care about errors in unprepare (may be the old connection went bad)
@@ -523,13 +516,13 @@ namespace Microsoft.Data.SqlClient
                         finally
                         {
                             // clean prepare status (even successful Unprepare does not do that)
-                            _prepareHandle = -1;
+                            _prepareHandle = s_cachedInvalidPrepareHandle;
                             _execType = EXECTYPE.UNPREPARED;
                         }
                     }
                 }
                 _activeConnection = value;
-                SqlClientEventSource.Log.TryTraceEvent("<sc.SqlCommand.set_Connection|API> {0}, {1}", ObjectID, value?.ObjectID);
+                SqlClientEventSource.Log.TryTraceEvent("SqlCommand.Set_Connection | API | ObjectId {0}, Client Connection Id {1}", ObjectID, value?.ClientConnectionId);
             }
         }
 
@@ -575,22 +568,6 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-        /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/NotificationAutoEnlist/*'/>
-        [DefaultValue(true)]
-        [ResCategory(StringsHelper.ResourceNames.DataCategory_Notification)]
-        [ResDescription(StringsHelper.ResourceNames.SqlCommand_NotificationAutoEnlist)]
-        public bool NotificationAutoEnlist
-        {
-            get
-            {
-                return _notificationAutoEnlist;
-            }
-            set
-            {
-                _notificationAutoEnlist = value;
-            }
-        }
-
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/Notification/*'/>
         [Browsable(false)]
         [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)] // MDAC 90471
@@ -604,9 +581,9 @@ namespace Microsoft.Data.SqlClient
             }
             set
             {
-                SqlClientEventSource.Log.TryTraceEvent("<sc.SqlCommand.set_Notification|API> {0}", ObjectID);
                 _sqlDep = null;
                 _notification = value;
+                SqlClientEventSource.Log.TryTraceEvent("SqlCommand.Set_Notification | API | Object Id {0}", ObjectID);
             }
         }
 
@@ -616,7 +593,8 @@ namespace Microsoft.Data.SqlClient
             {
                 if (_activeConnection != null)
                 {
-                    if (_activeConnection.StatisticsEnabled)
+                    if (_activeConnection.StatisticsEnabled ||
+                        s_diagnosticListener.IsEnabled(SqlClientCommandAfter.Name))
                     {
                         return _activeConnection.Statistics;
                     }
@@ -710,7 +688,7 @@ namespace Microsoft.Data.SqlClient
             {
                 if (value < 0)
                 {
-                    throw ADP.InvalidCommandTimeout(value, nameof(CommandTimeout));
+                    throw ADP.InvalidCommandTimeout(value);
                 }
 
                 if (value != _commandTimeout)
@@ -793,7 +771,6 @@ namespace Microsoft.Data.SqlClient
             set
             {
                 _designTimeInvisible = !value;
-                TypeDescriptor.Refresh(this);
             }
         }
 
@@ -888,8 +865,6 @@ namespace Microsoft.Data.SqlClient
                         {
                             throw;
                         }
-
-                        ADP.TraceExceptionWithoutRethrow(e);
                     }
                 }
             }
@@ -904,73 +879,73 @@ namespace Microsoft.Data.SqlClient
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/Prepare/*'/>
         public override void Prepare()
         {
-            SqlConnection.ExecutePermission.Demand();
-
             // Reset _pendingCancel upon entry into any Execute - used to synchronize state
             // between entry into Execute* API and the thread obtaining the stateObject.
             _pendingCancel = false;
 
+            SqlStatistics statistics = null;
             using (TryEventScope.Create("SqlCommand.Prepare | API | Object Id {0}", ObjectID))
             {
-                SqlClientEventSource.Log.TryCorrelationTraceEvent("<sc.SqlCommand.Prepare|API|Correlation> ObjectID {0}, ActivityID {1}", ObjectID, ActivityCorrelator.Current);
-
-                SqlStatistics statistics = SqlStatistics.StartTimer(Statistics);
-
-                // only prepare if batch with parameters
-                // MDAC BUG #'s 73776 & 72101
-                if (
-                    IsPrepared && !IsDirty
-                    || CommandType == CommandType.StoredProcedure
-                    || (CommandType == CommandType.Text && GetParameterCount(_parameters) == 0)
-                )
+                SqlClientEventSource.Log.TryCorrelationTraceEvent("SqlCommand.Prepare | API | Correlation | Object Id {0}, ActivityID {1}, Client Connection Id {2}", ObjectID, ActivityCorrelator.Current, Connection?.ClientConnectionId);
+                try
                 {
-                    if (Statistics != null)
-                    {
-                        Statistics.SafeIncrement(ref Statistics._prepares);
-                    }
-                    _hiddenPrepare = false;
-                }
-                else
-                {
-                    // Validate the command outside of the try\catch to avoid putting the _stateObj on error
-                    ValidateCommand(isAsync: false);
+                    statistics = SqlStatistics.StartTimer(Statistics);
 
-                    bool processFinallyBlock = true;
-                    try
+                    // only prepare if batch with parameters
+                    if (this.IsPrepared && !this.IsDirty
+                        || (this.CommandType == CommandType.StoredProcedure)
+                        || ((System.Data.CommandType.Text == this.CommandType)
+                                && (0 == GetParameterCount(_parameters))))
                     {
-                        // NOTE: The state object isn't actually needed for this, but it is still here for back-compat (since it does a bunch of checks)
-                        GetStateObject();
-
-                        // Loop through parameters ensuring that we do not have unspecified types, sizes, scales, or precisions
-                        if (_parameters != null)
+                        if (Statistics != null)
                         {
-                            int count = _parameters.Count;
-                            for (int i = 0; i < count; ++i)
+                            Statistics.SafeIncrement(ref Statistics._prepares);
+                        }
+                        _hiddenPrepare = false;
+                    }
+                    else
+                    {
+                        // Validate the command outside of the try/catch to avoid putting the _stateObj on error
+                        ValidateCommand(isAsync: false);
+
+                        bool processFinallyBlock = true;
+                        try
+                        {
+                            // NOTE: The state object isn't actually needed for this, but it is still here for back-compat (since it does a bunch of checks)
+                            GetStateObject();
+
+                            // Loop through parameters ensuring that we do not have unspecified types, sizes, scales, or precisions
+                            if (_parameters != null)
                             {
-                                _parameters[i].Prepare(this); // MDAC 67063
+                                int count = _parameters.Count;
+                                for (int i = 0; i < count; ++i)
+                                {
+                                    _parameters[i].Prepare(this);
+                                }
+                            }
+
+                            InternalPrepare();
+                        }
+                        catch (Exception e)
+                        {
+                            processFinallyBlock = ADP.IsCatchableExceptionType(e);
+                            throw;
+                        }
+                        finally
+                        {
+                            if (processFinallyBlock)
+                            {
+                                _hiddenPrepare = false; // The command is now officially prepared
+
+                                ReliablePutStateObject();
                             }
                         }
-
-                        InternalPrepare();
-                    }
-                    // @TODO: CER Exception Handling was removed here (see GH#3581)
-                    catch (Exception e)
-                    {
-                        processFinallyBlock = ADP.IsCatchableExceptionType(e);
-                        throw;
-                    }
-                    finally
-                    {
-                        if (processFinallyBlock)
-                        {
-                            _hiddenPrepare = false; // The command is now officially prepared
-
-                            ReliablePutStateObject();
-                        }
                     }
                 }
-
-                SqlStatistics.StopTimer(statistics);
+                finally
+                {
+                    SqlStatistics.StopTimer(statistics);
+                }
             }
         }
 
@@ -1019,7 +994,7 @@ namespace Microsoft.Data.SqlClient
             if ((_activeConnection.CloseCount != _preparedConnectionCloseCount) || (_activeConnection.ReconnectCount != _preparedConnectionReconnectCount))
             {
                 // reset our handle
-                _prepareHandle = -1;
+                _prepareHandle = s_cachedInvalidPrepareHandle;
             }
 
             _cachedMetaData = null;
@@ -1028,7 +1003,7 @@ namespace Microsoft.Data.SqlClient
 
         // Cancel is supposed to be multi-thread safe.
         // It doesn't make sense to verify the connection exists or that it is open during cancel
-        // because immediately after checkin the connection can be closed or removed via another thread.
+        // because immediately after checking the connection can be closed or removed via another thread.
         //
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/Cancel/*'/>
         public override void Cancel()
@@ -1086,8 +1061,8 @@ namespace Microsoft.Data.SqlClient
                         }
 
                         if (!_pendingCancel)
-                        {
-                            // Do nothing if aleady pending.
+                        { 
+                            // Do nothing if already pending.
                           // Before attempting actual cancel, set the _pendingCancel flag to false.
                           // This denotes to other thread before obtaining stateObject from the
                           // session pool that there is another thread wishing to cancel.
@@ -1109,7 +1084,6 @@ namespace Microsoft.Data.SqlClient
                                 }
                             }
                         }
-                        // @TODO: CER Exception Handling was removed here (see GH#3581)
                     }
                 }
                 finally
@@ -1160,15 +1134,14 @@ namespace Microsoft.Data.SqlClient
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/ExecuteScalar/*'/>
         public override object ExecuteScalar()
         {
-            SqlConnection.ExecutePermission.Demand();
-
             // Reset _pendingCancel upon entry into any Execute - used to synchronize state
             // between entry into Execute* API and the thread obtaining the stateObject.
             _pendingCancel = false;
-            SqlStatistics statistics = null;
 
+            using (DiagnosticScope diagnosticScope = s_diagnosticListener.CreateCommandScope(this, _transaction))
             using (TryEventScope.Create("SqlCommand.ExecuteScalar | API | ObjectId {0}", ObjectID))
             {
+                SqlStatistics statistics = null;
                 bool success = false;
                 int? sqlExceptionNumber = null;
                 SqlClientEventSource.Log.TryCorrelationTraceEvent("SqlCommand.ExecuteScalar | API | Correlation | Object Id {0}, Activity Id {1}, Client Connection Id {2}, Command Text '{3}'", ObjectID, ActivityCorrelator.Current, Connection?.ClientConnectionId, CommandText);
@@ -1180,13 +1153,16 @@ namespace Microsoft.Data.SqlClient
                     SqlDataReader ds = IsProviderRetriable
                         ? RunExecuteReaderWithRetry(0, RunBehavior.ReturnImmediately, returnStream: true)
                         : RunExecuteReader(0, RunBehavior.ReturnImmediately, returnStream: true);
-                    object result = CompleteExecuteScalar(ds, _batchRPCMode);
                     success = true;
-                    return result;
+                    return CompleteExecuteScalar(ds, _batchRPCMode);
                 }
-                catch (SqlException ex)
+                catch (Exception ex)
                 {
-                    sqlExceptionNumber = ex.Number;
+                    diagnosticScope.SetException(ex);
+                    if (ex is SqlException sqlException)
+                    {
+                        sqlExceptionNumber = sqlException.Number;
+                    }
                     throw;
                 }
                 finally
@@ -1249,16 +1225,14 @@ namespace Microsoft.Data.SqlClient
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/ExecuteNonQuery[@name="default"]/*'/>
         public override int ExecuteNonQuery()
         {
-            SqlConnection.ExecutePermission.Demand();
-
             // Reset _pendingCancel upon entry into any Execute - used to synchronize state
             // between entry into Execute* API and the thread obtaining the stateObject.
             _pendingCancel = false;
 
-            SqlStatistics statistics = null;
-
-            using (TryEventScope.Create("<sc.SqlCommand.ExecuteNonQuery|API> {0}", ObjectID))
+            using (var diagnosticScope = s_diagnosticListener.CreateCommandScope(this, _transaction))
+            using (TryEventScope.Create("SqlCommand.ExecuteNonQuery | API | Object Id {0}", ObjectID))
             {
+                SqlStatistics statistics = null;
                 bool success = false;
                 int? sqlExceptionNumber = null;
                 SqlClientEventSource.Log.TryCorrelationTraceEvent("SqlCommand.ExecuteNonQuery | API | Correlation | Object Id {0}, ActivityID {1}, Client Connection Id {2}, Command Text {3}", ObjectID, ActivityCorrelator.Current, Connection?.ClientConnectionId, CommandText);
@@ -1271,7 +1245,7 @@ namespace Microsoft.Data.SqlClient
                     {
                         InternalExecuteNonQueryWithRetry(
                             sendToPipe: false,
-                            CommandTimeout,
+                            timeout: CommandTimeout,
                             out _,
                             asyncWrite: false,
                             isRetry: false);
@@ -1283,9 +1257,13 @@ namespace Microsoft.Data.SqlClient
                     success = true;
                     return _rowsAffected;
                 }
-                catch (SqlException ex)
+                catch (Exception ex)
                 {
-                    sqlExceptionNumber = ex.Number;
+                    diagnosticScope.SetException(ex);
+                    if (ex is SqlException sqlException)
+                    {
+                        sqlExceptionNumber = sqlException.Number;
+                    }
                     throw;
                 }
                 finally
@@ -1297,16 +1275,13 @@ namespace Microsoft.Data.SqlClient
         }
 
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/BeginExecuteNonQuery[@name="default"]/*'/>
-        [System.Security.Permissions.HostProtectionAttribute(ExternalThreading = true)]
         public IAsyncResult BeginExecuteNonQuery() =>
             BeginExecuteNonQuery(null, null);
 
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/BeginExecuteNonQuery[@name="AsyncCallbackAndStateObject"]/*'/>
-        [System.Security.Permissions.HostProtectionAttribute(ExternalThreading = true)]
         public IAsyncResult BeginExecuteNonQuery(AsyncCallback callback, object stateObject)
         {
-            SqlClientEventSource.Log.TryCorrelationTraceEvent("<sc.SqlCommand.BeginExecuteNonQuery|API|Correlation> ObjectID {0}, ActivityID {1}", ObjectID, ActivityCorrelator.Current);
-            SqlConnection.ExecutePermission.Demand();
+            SqlClientEventSource.Log.TryCorrelationTraceEvent("SqlCommand.BeginExecuteNonQuery | API | Correlation | Object Id {0}, Activity Id {1}, Client Connection Id {2}, Command Text '{3}'", ObjectID, ActivityCorrelator.Current, Connection?.ClientConnectionId, CommandText);
             return BeginExecuteNonQueryInternal(0, callback, stateObject, 0, isRetry: false);
         }
 
@@ -1337,7 +1312,6 @@ namespace Microsoft.Data.SqlClient
                 if (!isRetry)
                 {
                     statistics = SqlStatistics.StartTimer(Statistics);
-                    WriteBeginExecuteEvent();
                 }
 
                 bool usedCache;
@@ -1355,7 +1329,14 @@ namespace Microsoft.Data.SqlClient
 
                     if (execNQ != null)
                     {
-                        AsyncHelper.ContinueTaskWithState(execNQ, localCompletion, this, (object state) => ((SqlCommand)state).BeginExecuteNonQueryInternalReadStage(localCompletion));
+                        AsyncHelper.ContinueTaskWithState(execNQ, localCompletion,
+                            state: Tuple.Create(this, localCompletion),
+                            onSuccess: state =>
+                            {
+                                var parameters = (Tuple<SqlCommand, TaskCompletionSource<object>>)state;
+                                parameters.Item1.BeginExecuteNonQueryInternalReadStage(parameters.Item2);
+                            }
+                        );
                     }
                     else
                     {
@@ -1391,9 +1372,9 @@ namespace Microsoft.Data.SqlClient
                         {
                             return command.InternalEndExecuteNonQuery(asyncResult, isInternal, endMethod);
                         },
-                        retryFunc: static (SqlCommand command, CommandBehavior commandBehavior, AsyncCallback callback, object stateObject, int timeout, bool isRetry, bool asyncWrite) =>
+                        retryFunc: static (SqlCommand command, CommandBehavior behavior, AsyncCallback callback, object stateObject, int timeout, bool isRetry, bool asyncWrite) =>
                         {
-                            return command.BeginExecuteNonQueryInternal(commandBehavior, callback, stateObject, timeout, isRetry, asyncWrite);
+                            return command.BeginExecuteNonQueryInternal(behavior, callback, stateObject, timeout, isRetry, asyncWrite);
                         },
                         nameof(EndExecuteNonQuery)))
                 {
@@ -1403,7 +1384,10 @@ namespace Microsoft.Data.SqlClient
                 // Add callback after work is done to avoid overlapping Begin/End methods
                 if (callback != null)
                 {
-                    globalCompletion.Task.ContinueWith((t) => callback(t), TaskScheduler.Default);
+                    globalCompletion.Task.ContinueWith(
+                        static (Task<object> task, object state) => ((AsyncCallback)state)(task),
+                        state: callback
+                    );
                 }
 
                 return globalCompletion.Task;
@@ -1428,10 +1412,7 @@ namespace Microsoft.Data.SqlClient
             {
                 // Similarly, if an exception occurs put the stateObj back into the pool.
                 // and reset async cache information to allow a second async execute
-                if (CachedAsyncState != null)
-                {
-                    CachedAsyncState.ResetAsyncState();
-                }
+                CachedAsyncState?.ResetAsyncState();
                 ReliablePutStateObject();
                 throw;
             }
@@ -1533,7 +1514,7 @@ namespace Microsoft.Data.SqlClient
             if (_stateObj == null)
             {
                 var reconnectionCompletionSource = _reconnectionCompletionSource;
-                if (reconnectionCompletionSource != null && reconnectionCompletionSource.Task.IsCanceled)
+                if (reconnectionCompletionSource != null && reconnectionCompletionSource.Task != null && reconnectionCompletionSource.Task.IsCanceled)
                 {
                     throw SQL.CR_ReconnectionCancelled();
                 }
@@ -1618,7 +1599,6 @@ namespace Microsoft.Data.SqlClient
             SqlClientEventSource.Log.TryTraceEvent("SqlCommand.InternalEndExecuteNonQuery | INFO | ObjectId {0}, Client Connection Id {1}, MARS={2}, AsyncCommandInProgress={3}",
                                                     _activeConnection?.ObjectID, _activeConnection?.ClientConnectionId,
                                                     _activeConnection?.Parser?.MARSOn, _activeConnection?.AsyncCommandInProgress);
-
             VerifyEndExecuteState((Task)asyncResult, endMethod);
             WaitForAsyncResults(asyncResult, isInternal);
 
@@ -1697,7 +1677,6 @@ namespace Microsoft.Data.SqlClient
 
             Debug.Assert(_stateObj == null, "non-null state object in EndExecuteNonQuery");
             return _rowsAffected;
-            // @TODO: CER Exception Handling was removed here (see GH#3581)
         }
 
         private Task InternalExecuteNonQuery(
@@ -1717,21 +1696,23 @@ namespace Microsoft.Data.SqlClient
             SqlStatistics statistics = Statistics;
             _rowsAffected = -1;
 
-            // @devnote: this function may throw for an invalid connection
-            // @devnote: returns false for empty command text
+            // this function may throw for an invalid connection
+            // returns false for empty command text
             if (!isRetry)
             {
                 ValidateCommand(isAsync, methodName);
             }
+
             CheckNotificationStateAndAutoEnlist(); // Only call after validate - requires non null connection!
 
             Task task = null;
 
             // Always Encrypted generally operates only on parameterized queries. However enclave based Always encrypted also supports unparameterized queries
             // We skip this block for enclave based always encrypted so that we can make a call to SQL Server to get the encryption information
-            if (!ShouldUseEnclaveBasedWorkflow && !_batchRPCMode && CommandType == CommandType.Text && GetParameterCount(_parameters) == 0)
+            if (!ShouldUseEnclaveBasedWorkflow && !_batchRPCMode && (CommandType.Text == CommandType) && (0 == GetParameterCount(_parameters)))
             {
-                Debug.Assert(!sendToPipe, "trying to send non-context command to pipe");
+                Debug.Assert(!sendToPipe, "Trying to send non-context command to pipe");
+
                 if (statistics != null)
                 {
                     if (!IsDirty && IsPrepared)
@@ -1746,6 +1727,7 @@ namespace Microsoft.Data.SqlClient
 
                 // We should never get here for a retry since we only have retries for parameters.
                 Debug.Assert(!isRetry);
+                SqlClientEventSource.Log.TryTraceEvent("SqlCommand.InternalExecuteNonQuery | INFO | Object Id {0}, RPC execute method name {1}, isAsync {2}, isRetry {3}", ObjectID, methodName, isAsync, isRetry);
 
                 task = RunExecuteNonQueryTds(methodName, isAsync, timeout, asyncWrite);
             }
@@ -1766,12 +1748,15 @@ namespace Microsoft.Data.SqlClient
                     asyncWrite,
                     isRetry,
                     methodName);
-
+                
                 if (reader != null)
                 {
                     if (task != null)
                     {
-                        task = AsyncHelper.CreateContinuationTask(task, () => reader.Close());
+                        task = AsyncHelper.CreateContinuationTaskWithState(task,
+                            state: reader,
+                            onSuccess: state => ((SqlDataReader)state).Close()
+                        );
                     }
                     else
                     {
@@ -1779,24 +1764,22 @@ namespace Microsoft.Data.SqlClient
                     }
                 }
             }
+
             Debug.Assert(isAsync || _stateObj == null, "non-null state object in InternalExecuteNonQuery");
             return task;
-            // @TODO: CER Exception Handling was removed here (see GH#3581)
         }
 
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/ExecuteXmlReader/*'/>
         public XmlReader ExecuteXmlReader()
         {
-            SqlConnection.ExecutePermission.Demand();
-
             // Reset _pendingCancel upon entry into any Execute - used to synchronize state
             // between entry into Execute* API and the thread obtaining the stateObject.
             _pendingCancel = false;
 
-            SqlStatistics statistics = null;
-
+            using (DiagnosticScope diagnosticScope = s_diagnosticListener.CreateCommandScope(this, _transaction))
             using (TryEventScope.Create("SqlCommand.ExecuteXmlReader | API | Object Id {0}", ObjectID))
             {
+                SqlStatistics statistics = null;
                 bool success = false;
                 int? sqlExceptionNumber = null;
                 SqlClientEventSource.Log.TryCorrelationTraceEvent("SqlCommand.ExecuteXmlReader | API | Correlation | Object Id {0}, Activity Id {1}, Client Connection Id {2}, Command Text '{3}'", ObjectID, ActivityCorrelator.Current, Connection?.ClientConnectionId, CommandText);
@@ -1807,16 +1790,19 @@ namespace Microsoft.Data.SqlClient
                     WriteBeginExecuteEvent();
 
                     // use the reader to consume metadata
-                    SqlDataReader ds = IsProviderRetriable
-                        ? RunExecuteReaderWithRetry(CommandBehavior.SequentialAccess, RunBehavior.ReturnImmediately, returnStream: true)
-                        : RunExecuteReader(CommandBehavior.SequentialAccess, RunBehavior.ReturnImmediately, returnStream: true);
-                    XmlReader result = CompleteXmlReader(ds);
+                    SqlDataReader ds = IsProviderRetriable ?
+                        RunExecuteReaderWithRetry(CommandBehavior.SequentialAccess, RunBehavior.ReturnImmediately, returnStream: true) :
+                        RunExecuteReader(CommandBehavior.SequentialAccess, RunBehavior.ReturnImmediately, returnStream: true);
                     success = true;
-                    return result;
+                    return CompleteXmlReader(ds);
                 }
-                catch (SqlException ex)
+                catch (Exception ex)
                 {
-                    sqlExceptionNumber = ex.Number;
+                    diagnosticScope.SetException(ex);
+                    if (ex is SqlException sqlException)
+                    {
+                        sqlExceptionNumber = sqlException.Number;
+                    }
                     throw;
                 }
                 finally
@@ -1828,7 +1814,6 @@ namespace Microsoft.Data.SqlClient
         }
 
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/BeginExecuteXmlReader[@name="default"]/*'/>
-        [System.Security.Permissions.HostProtectionAttribute(ExternalThreading = true)]
         public IAsyncResult BeginExecuteXmlReader()
         {
             // BeginExecuteXmlReader will track executiontime
@@ -1836,11 +1821,9 @@ namespace Microsoft.Data.SqlClient
         }
 
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/BeginExecuteXmlReader[@name="AsyncCallbackAndstateObject"]/*'/>
-        [System.Security.Permissions.HostProtectionAttribute(ExternalThreading = true)]
         public IAsyncResult BeginExecuteXmlReader(AsyncCallback callback, object stateObject)
         {
             SqlClientEventSource.Log.TryCorrelationTraceEvent("SqlCommand.BeginExecuteXmlReader | API | Correlation | Object Id {0}, Activity Id {1}, Client Connection Id {2}, Command Text '{3}'", ObjectID, ActivityCorrelator.Current, Connection?.ClientConnectionId, CommandText);
-            SqlConnection.ExecutePermission.Demand();
             return BeginExecuteXmlReaderInternal(CommandBehavior.SequentialAccess, callback, stateObject, 0, isRetry: false);
         }
 
@@ -1905,7 +1888,14 @@ namespace Microsoft.Data.SqlClient
 
                 if (writeTask != null)
                 {
-                    AsyncHelper.ContinueTaskWithState(writeTask, localCompletion, this, (object state) => ((SqlCommand)state).BeginExecuteXmlReaderInternalReadStage(localCompletion));
+                    AsyncHelper.ContinueTaskWithState(writeTask, localCompletion,
+                        state: Tuple.Create(this, localCompletion),
+                        onSuccess: state =>
+                        {
+                            var parameters = (Tuple<SqlCommand, TaskCompletionSource<object>>)state;
+                            parameters.Item1.BeginExecuteXmlReaderInternalReadStage(parameters.Item2);
+                        }
+                    );
                 }
                 else
                 {
@@ -1930,7 +1920,7 @@ namespace Microsoft.Data.SqlClient
                         },
                         retryFunc: static (SqlCommand command, CommandBehavior behavior, AsyncCallback callback, object stateObject, int timeout, bool isRetry, bool asyncWrite) =>
                         {
-                            return command.BeginExecuteXmlReaderInternal(behavior, callback, stateObject, timeout, isRetry, asyncWrite);                    
+                            return command.BeginExecuteXmlReaderInternal(behavior, callback, stateObject, timeout, isRetry, asyncWrite);
                         },
                         endMethod: nameof(EndExecuteXmlReader)))
                 {
@@ -1940,9 +1930,12 @@ namespace Microsoft.Data.SqlClient
                 // Add callback after work is done to avoid overlapping Begin/End methods
                 if (callback != null)
                 {
-                    globalCompletion.Task.ContinueWith((t) => callback(t), TaskScheduler.Default);
+                    localCompletion.Task.ContinueWith(
+                        static (Task<object> task, object state) => ((AsyncCallback)state)(task),
+                        state: callback
+                    );
                 }
-                return globalCompletion.Task;
+                return localCompletion.Task;
             }
             finally
             {
@@ -2021,26 +2014,16 @@ namespace Microsoft.Data.SqlClient
             int? sqlExceptionNumber = null;
             try
             {
-                XmlReader result = CompleteXmlReader(
-                    InternalEndExecuteReader(asyncResult, isInternal: false,  nameof(EndExecuteXmlReader)),
-                    isAsync: true);
                 success = true;
-                return result;
-            }
-            catch (SqlException e)
-            {
-                sqlExceptionNumber = e.Number;
-                if (CachedAsyncState != null)
-                {
-                    CachedAsyncState.ResetAsyncState();
-                };
-
-                //  SqlException is always catchable 
-                ReliablePutStateObject();
-                throw;
+                return CompleteXmlReader(InternalEndExecuteReader(asyncResult, false, nameof(EndExecuteXmlReader)), true);
             }
             catch (Exception e)
             {
+                if (e is SqlException)
+                {
+                    SqlException ex = (SqlException)e;
+                    sqlExceptionNumber = ex.Number;
+                }
                 if (CachedAsyncState != null)
                 {
                     CachedAsyncState.ResetAsyncState();
@@ -2091,26 +2074,21 @@ namespace Microsoft.Data.SqlClient
         }
 
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/BeginExecuteXmlReader[@name="default"]/*'/>
-        [HostProtection(ExternalThreading = true)]
         public IAsyncResult BeginExecuteReader() =>
             BeginExecuteReader(callback: null, stateObject: null, CommandBehavior.Default);
 
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/BeginExecuteXmlReader[@name="AsyncCallbackAndstateObject"]/*'/>
-        [HostProtection(ExternalThreading = true)]
         public IAsyncResult BeginExecuteReader(AsyncCallback callback, object stateObject) =>
             BeginExecuteReader(callback, stateObject, CommandBehavior.Default);
 
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/BeginExecuteReader[@name="CommandBehavior"]/*'/>
-        [HostProtection(ExternalThreading = true)]
         public IAsyncResult BeginExecuteReader(CommandBehavior behavior) =>
             BeginExecuteReader(callback: null, stateObject: null, behavior);
 
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/BeginExecuteReader[@name="AsyncCallbackAndstateObjectAndCommandBehavior"]/*'/>
-        [HostProtection(ExternalThreading = true)]
         public IAsyncResult BeginExecuteReader(AsyncCallback callback, object stateObject, CommandBehavior behavior)
         {
             SqlClientEventSource.Log.TryCorrelationTraceEvent("SqlCommand.BeginExecuteReader | API | Correlation | Object Id {0}, Behavior {1}, Activity Id {2}, Client Connection Id {3}, Command Text '{4}'", ObjectID, (int)behavior, ActivityCorrelator.Current, Connection?.ClientConnectionId, CommandText);
-            SqlConnection.ExecutePermission.Demand();
             return BeginExecuteReaderInternal(behavior, callback, stateObject, 0, isRetry: false);
         }
 
@@ -2140,8 +2118,6 @@ namespace Microsoft.Data.SqlClient
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/ExecuteReader[@name="CommandBehavior"]/*'/>
         new public SqlDataReader ExecuteReader(CommandBehavior behavior)
         {
-            SqlConnection.ExecutePermission.Demand();
-
             // Reset _pendingCancel upon entry into any Execute - used to synchronize state
             // between entry into Execute* API and the thread obtaining the stateObject.
             _pendingCancel = false;
@@ -2149,6 +2125,8 @@ namespace Microsoft.Data.SqlClient
             SqlStatistics statistics = null;
             bool success = false;
             int? sqlExceptionNumber = null;
+            Exception e = null;
+            Guid operationId = s_diagnosticListener.WriteCommandBefore(this, _transaction);
 
             using (TryEventScope.Create("SqlCommand.ExecuteReader | API | Object Id {0}", ObjectID))
             {
@@ -2156,22 +2134,32 @@ namespace Microsoft.Data.SqlClient
                 {
                     WriteBeginExecuteEvent();
                     statistics = SqlStatistics.StartTimer(Statistics);
-                    SqlDataReader result = IsProviderRetriable ?
+                    return IsProviderRetriable ?
                         RunExecuteReaderWithRetry(behavior, RunBehavior.ReturnImmediately, returnStream: true) :
-                        RunExecuteReader(behavior, RunBehavior.ReturnImmediately, true);
-                    success = true;
-                    return result;
+                        RunExecuteReader(behavior, RunBehavior.ReturnImmediately, returnStream: true);
                 }
-                catch (SqlException e)
+                catch (Exception ex)
                 {
-                    sqlExceptionNumber = e.Number;
+                    if (ex is SqlException sqlException)
+                    {
+                        sqlExceptionNumber = sqlException.Number;
+                    }
+
+                    e = ex;
                     throw;
                 }
-                // @TODO: CER Exception Handling was removed here (see GH#3581)
                 finally
                 {
                     SqlStatistics.StopTimer(statistics);
                     WriteEndExecuteEvent(success, sqlExceptionNumber, synchronous: true);
+                    if (e != null)
+                    {
+                        s_diagnosticListener.WriteCommandError(operationId, this, _transaction, e);
+                    }
+                    else
+                    {
+                        s_diagnosticListener.WriteCommandAfter(operationId, this, _transaction);
+                    }
                 }
             }
         }
@@ -2229,28 +2217,18 @@ namespace Microsoft.Data.SqlClient
             int? sqlExceptionNumber = null;
             try
             {
-                statistics = SqlStatistics.StartTimer(Statistics);
-                SqlDataReader result = InternalEndExecuteReader(
-                    asyncResult,
-                    isInternal: false,
-                    nameof(EndExecuteReader));
                 success = true;
-                return result;
-            }
-            catch (SqlException e)
-            {
-                sqlExceptionNumber = e.Number;
-                if (CachedAsyncState != null)
-                {
-                    CachedAsyncState.ResetAsyncState();
-                };
-
-                //  SqlException is always catchable 
-                ReliablePutStateObject();
-                throw;
+                statistics = SqlStatistics.StartTimer(Statistics);
+                return InternalEndExecuteReader(asyncResult, false, nameof(EndExecuteReader));
             }
             catch (Exception e)
             {
+                if (e is SqlException)
+                {
+                    SqlException exception = (SqlException)e;
+                    sqlExceptionNumber = exception.Number;
+                }
+
                 if (CachedAsyncState != null)
                 {
                     CachedAsyncState.ResetAsyncState();
@@ -2273,10 +2251,18 @@ namespace Microsoft.Data.SqlClient
             if (task.IsFaulted)
             {
                 Exception e = task.Exception.InnerException;
+                if (!_parentOperationStarted)
+                {
+                    s_diagnosticListener.WriteCommandError(operationId, this, _transaction, e);
+                }
                 source.SetException(e);
             }
             else
             {
+                if (!_parentOperationStarted)
+                {
+                    s_diagnosticListener.WriteCommandAfter(operationId, this, _transaction);
+                }
                 if (task.IsCanceled)
                 {
                     source.SetCanceled();
@@ -2286,11 +2272,6 @@ namespace Microsoft.Data.SqlClient
                     source.SetResult(task.Result);
                 }
             }
-        }
-
-        private IAsyncResult BeginExecuteReaderAsync(CommandBehavior behavior, AsyncCallback callback, object stateObject)
-        {
-            return BeginExecuteReaderInternal(behavior, callback, stateObject, CommandTimeout, isRetry: false, asyncWrite: true);
         }
 
         private IAsyncResult BeginExecuteReaderInternal(CommandBehavior behavior, AsyncCallback callback, object stateObject, int timeout, bool isRetry, bool asyncWrite = false)
@@ -2344,12 +2325,22 @@ namespace Microsoft.Data.SqlClient
 
                     // For async, RunExecuteReader will never put the stateObj back into the pool, so do so now.
                     ReliablePutStateObject();
-                    throw;
+                    if (isRetry || e is not EnclaveDelegate.RetryableEnclaveQueryExecutionException)
+                    {
+                        throw;
+                    }
                 }
 
                 if (writeTask != null)
                 {
-                    AsyncHelper.ContinueTaskWithState(writeTask, localCompletion, this, (object state) => ((SqlCommand)state).BeginExecuteReaderInternalReadStage(localCompletion));
+                    AsyncHelper.ContinueTaskWithState(writeTask, localCompletion,
+                        state: Tuple.Create(this, localCompletion),
+                        onSuccess: state =>
+                        {
+                            var parameters = (Tuple<SqlCommand, TaskCompletionSource<object>>)state;
+                            parameters.Item1.BeginExecuteReaderInternalReadStage(parameters.Item2);
+                        }
+                    );
                 }
                 else
                 {
@@ -2384,7 +2375,10 @@ namespace Microsoft.Data.SqlClient
                 // Add callback after work is done to avoid overlapping Begin/End methods
                 if (callback != null)
                 {
-                    globalCompletion.Task.ContinueWith((t) => callback(t), TaskScheduler.Default);
+                    globalCompletion.Task.ContinueWith(
+                        static (Task<object> task, object state) => ((AsyncCallback)state)(task),
+                        state: callback
+                    );
                 }
 
                 return globalCompletion.Task;
@@ -2421,125 +2415,18 @@ namespace Microsoft.Data.SqlClient
             {
                 long firstAttemptStart = ADP.TimerCurrent();
 
-                localCompletion.Task.ContinueWith(tsk =>
-                {
-                    if (tsk.IsFaulted)
-                    {
-                        globalCompletion.TrySetException(tsk.Exception.InnerException);
-                    }
-                    else if (tsk.IsCanceled)
-                    {
-                        globalCompletion.TrySetCanceled();
-                    }
-                    else
-                    {
-                        try
-                        {
-                            // Mark that we initiated the internal EndExecute. This should always be false until we set it here.
-                            Debug.Assert(!_internalEndExecuteInitiated);
-                            _internalEndExecuteInitiated = true;
-
-                            // lock on _stateObj prevents races with close/cancel.
-                            lock (_stateObj)
-                            {
-                                endFunc(this, tsk, /*isInternal:*/ true, endMethod);
-                            }
-                            globalCompletion.TrySetResult(tsk.Result);
-                        }
-                        catch (Exception e)
-                        {
-                            // Put the state object back to the cache.
-                            // Do not reset the async state, since this is managed by the user Begin/End and not internally.
-                            if (ADP.IsCatchableExceptionType(e))
-                            {
-                                ReliablePutStateObject();
-                            }
-
-                            bool shouldRetry = e is EnclaveDelegate.RetryableEnclaveQueryExecutionException;
-
-                            // Check if we have an error indicating that we can retry.
-                            if (e is SqlException)
-                            {
-                                SqlException sqlEx = e as SqlException;
-
-                                for (int i = 0; i < sqlEx.Errors.Count; i++)
-                                {
-                                    if ((usedCache && (sqlEx.Errors[i].Number == TdsEnums.TCE_CONVERSION_ERROR_CLIENT_RETRY)) ||
-                                         (ShouldUseEnclaveBasedWorkflow && (sqlEx.Errors[i].Number == TdsEnums.TCE_ENCLAVE_INVALID_SESSION_HANDLE)))
-                                    {
-                                        shouldRetry = true;
-                                        break;
-                                    }
-                                }
-                            }
-
-                            if (!shouldRetry)
-                            {
-                                // If we cannot retry, Reset the async state to make sure we leave a clean state.
-                                if (CachedAsyncState != null)
-                                {
-                                    CachedAsyncState.ResetAsyncState();
-                                }
-                                try
-                                {
-                                    _activeConnection.GetOpenTdsConnection().DecrementAsyncCount();
-
-                                    globalCompletion.TrySetException(e);
-                                }
-                                catch (Exception e2)
-                                {
-                                    globalCompletion.TrySetException(e2);
-                                }
-                            }
-                            else
-                            {
-                                // Remove the entry from the cache since it was inconsistent.
-                                SqlQueryMetadataCache.GetInstance().InvalidateCacheEntry(this);
-
-                                InvalidateEnclaveSession();
-
-                                try
-                                {
-                                    // Kick off the retry.
-                                    _internalEndExecuteInitiated = false;
-                                    Task<object> retryTask = (Task<object>)retryFunc(
-                                        this,
-                                        behavior,
-                                        null,
-                                        stateObject,
-                                        TdsParserStaticMethods.GetRemainingTimeout(timeout, firstAttemptStart),
-                                        /*isRetry:*/ true,
-                                        asyncWrite);
-
-                                    retryTask.ContinueWith(
-                                        static (Task<object> retryTask, object state) =>
-                                        {
-                                            TaskCompletionSource<object> completion = (TaskCompletionSource<object>)state;
-                                            if (retryTask.IsFaulted)
-                                            {
-                                                completion.TrySetException(retryTask.Exception.InnerException);
-                                            }
-                                            else if (retryTask.IsCanceled)
-                                            {
-                                                completion.TrySetCanceled();
-                                            }
-                                            else
-                                            {
-                                                completion.TrySetResult(retryTask.Result);
-                                            }
-                                        },
-                                        state: globalCompletion,
-                                        TaskScheduler.Default
-                                    );
-                                }
-                                catch (Exception e2)
-                                {
-                                    globalCompletion.TrySetException(e2);
-                                }
-                            }
-                        }
-                    }
-                }, TaskScheduler.Default);
+                CreateLocalCompletionTask(
+                    behavior,
+                    stateObject,
+                    timeout,
+                    usedCache,
+                    asyncWrite,
+                    globalCompletion,
+                    localCompletion,
+                    endFunc,
+                    retryFunc,
+                    endMethod,
+                    firstAttemptStart);
 
                 return true;
             }
@@ -2547,6 +2434,144 @@ namespace Microsoft.Data.SqlClient
             {
                 return false;
             }
+        }
+
+        private void CreateLocalCompletionTask(
+            CommandBehavior behavior,
+            object stateObject,
+            int timeout,
+            bool usedCache,
+            bool asyncWrite,
+            TaskCompletionSource<object> globalCompletion,
+            TaskCompletionSource<object> localCompletion,
+            Func<SqlCommand, IAsyncResult, bool, string, object> endFunc,
+            Func<SqlCommand, CommandBehavior, AsyncCallback, object, int, bool, bool, IAsyncResult> retryFunc,
+            string endMethod,
+            long firstAttemptStart
+        )
+        {
+            localCompletion.Task.ContinueWith(tsk =>
+            {
+                if (tsk.IsFaulted)
+                {
+                    globalCompletion.TrySetException(tsk.Exception.InnerException);
+                }
+                else if (tsk.IsCanceled)
+                {
+                    globalCompletion.TrySetCanceled();
+                }
+                else
+                {
+                    try
+                    {
+                        // Mark that we initiated the internal EndExecute. This should always be false until we set it here.
+                        Debug.Assert(!_internalEndExecuteInitiated);
+                        _internalEndExecuteInitiated = true;
+
+                        // lock on _stateObj prevents races with close/cancel.
+                        lock (_stateObj)
+                        {
+                            endFunc(this, tsk, /*isInternal:*/ true, endMethod);
+                        }
+
+                        globalCompletion.TrySetResult(tsk.Result);
+                    }
+                    catch (Exception e)
+                    {
+                        // Put the state object back to the cache.
+                        // Do not reset the async state, since this is managed by the user Begin/End and not internally.
+                        if (ADP.IsCatchableExceptionType(e))
+                        {
+                            ReliablePutStateObject();
+                        }
+
+                        bool shouldRetry = e is EnclaveDelegate.RetryableEnclaveQueryExecutionException;
+
+                        // Check if we have an error indicating that we can retry.
+                        if (e is SqlException)
+                        {
+                            SqlException sqlEx = e as SqlException;
+
+                            for (int i = 0; i < sqlEx.Errors.Count; i++)
+                            {
+                                if ((usedCache && (sqlEx.Errors[i].Number == TdsEnums.TCE_CONVERSION_ERROR_CLIENT_RETRY)) ||
+                                    (ShouldUseEnclaveBasedWorkflow &&
+                                     (sqlEx.Errors[i].Number == TdsEnums.TCE_ENCLAVE_INVALID_SESSION_HANDLE)))
+                                {
+                                    shouldRetry = true;
+                                    break;
+                                }
+                            }
+                        }
+
+                        if (!shouldRetry)
+                        {
+                            // If we cannot retry, Reset the async state to make sure we leave a clean state.
+                            if (CachedAsyncState != null)
+                            {
+                                CachedAsyncState.ResetAsyncState();
+                            }
+
+                            try
+                            {
+                                _activeConnection.GetOpenTdsConnection().DecrementAsyncCount();
+
+                                globalCompletion.TrySetException(e);
+                            }
+                            catch (Exception e2)
+                            {
+                                globalCompletion.TrySetException(e2);
+                            }
+                        }
+                        else
+                        {
+                            // Remove the entry from the cache since it was inconsistent.
+                            SqlQueryMetadataCache.GetInstance().InvalidateCacheEntry(this);
+
+                            InvalidateEnclaveSession();
+
+                            try
+                            {
+                                // Kick off the retry.
+                                _internalEndExecuteInitiated = false;
+                                Task<object> retryTask = (Task<object>)retryFunc(
+                                    this,
+                                    behavior,
+                                    null,
+                                    stateObject,
+                                    TdsParserStaticMethods.GetRemainingTimeout(timeout, firstAttemptStart),
+                                    /*isRetry:*/ true,
+                                    asyncWrite);
+
+                                retryTask.ContinueWith(
+                                    static (Task<object> retryTask, object state) =>
+                                    {
+                                        TaskCompletionSource<object> completion = (TaskCompletionSource<object>)state;
+                                        if (retryTask.IsFaulted)
+                                        {
+                                            completion.TrySetException(retryTask.Exception.InnerException);
+                                        }
+                                        else if (retryTask.IsCanceled)
+                                        {
+                                            completion.TrySetCanceled();
+                                        }
+                                        else
+                                        {
+                                            completion.TrySetResult(retryTask.Result);
+                                        }
+                                    }, 
+                                    state: globalCompletion,
+                                    TaskScheduler.Default
+                                );
+                            }
+                            catch (Exception e2)
+                            {
+                                globalCompletion.TrySetException(e2);
+                            }
+                        }
+                    }
+                }
+            }, TaskScheduler.Default);
         }
 
         private void InvalidateEnclaveSession()
@@ -2611,7 +2636,6 @@ namespace Microsoft.Data.SqlClient
             SqlDataReader reader = CompleteAsyncExecuteReader(isInternal);
             Debug.Assert(_stateObj == null, "non-null state object in InternalEndExecuteReader");
             return reader;
-            // @TODO: CER Exception Handling was removed here (see GH#3581)
         }
 
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/ExecuteNonQueryAsync[@name="CancellationToken"]/*'/>
@@ -2629,8 +2653,7 @@ namespace Microsoft.Data.SqlClient
         private Task<int> InternalExecuteNonQueryAsync(CancellationToken cancellationToken)
         {
             SqlClientEventSource.Log.TryCorrelationTraceEvent("SqlCommand.InternalExecuteNonQueryAsync | API | Correlation | Object Id {0}, Activity Id {1}, Client Connection Id {2}, Command Text '{3}'", ObjectID, ActivityCorrelator.Current, Connection?.ClientConnectionId, CommandText);
-            SqlConnection.ExecutePermission.Demand();
-            Guid operationId = Guid.Empty;
+            Guid operationId = s_diagnosticListener.WriteCommandBefore(this, _transaction);
 
             // connection can be used as state in RegisterForConnectionCloseNotification continuation
             // to avoid an allocation so use it as the state value if possible but it can be changed if
@@ -2649,26 +2672,25 @@ namespace Microsoft.Data.SqlClient
             }
 
             Task<int> returnedTask = source.Task;
+            returnedTask = RegisterForConnectionCloseNotification(returnedTask);
 
             ExecuteNonQueryAsyncCallContext context = new ExecuteNonQueryAsyncCallContext();
             context.Set(this, source, registration, operationId);
             try
             {
-                returnedTask = RegisterForConnectionCloseNotification(returnedTask);
-
                 Task<int>.Factory.FromAsync(
-                    beginMethod: static (AsyncCallback callback, object stateObject) =>
+                    beginMethod: static (AsyncCallback callback, object stateObject) => // with c# 10/NET6 add [StackTraceHidden] to this
                     {
                         return ((ExecuteNonQueryAsyncCallContext)stateObject).Command.BeginExecuteNonQueryAsync(callback, stateObject);
                     },
-                    endMethod: static (IAsyncResult asyncResult) =>
+                    endMethod: static (IAsyncResult asyncResult) => // with c# 10/NET6 add [StackTraceHidden] to this
                     {
                         return ((ExecuteNonQueryAsyncCallContext)asyncResult.AsyncState).Command.EndExecuteNonQueryAsync(asyncResult);
                     },
                     state: context
                 )
                 .ContinueWith(
-                    static (Task<int> task) =>
+                    static (Task<int> task) => // with c# 10/NET6 add [StackTraceHidden] to this
                     {
                         ExecuteNonQueryAsyncCallContext context = (ExecuteNonQueryAsyncCallContext)task.AsyncState;
                         SqlCommand command = context.Command;
@@ -2683,6 +2705,7 @@ namespace Microsoft.Data.SqlClient
             }
             catch (Exception e)
             {
+                s_diagnosticListener.WriteCommandError(operationId, this, _transaction, e);
                 source.SetException(e);
                 context.Dispose();
             }
@@ -2695,6 +2718,7 @@ namespace Microsoft.Data.SqlClient
             if (task.IsFaulted)
             {
                 Exception e = task.Exception.InnerException;
+                s_diagnosticListener.WriteCommandError(operationId, this, _transaction, e);
                 source.SetException(e);
             }
             else
@@ -2707,6 +2731,7 @@ namespace Microsoft.Data.SqlClient
                 {
                     source.SetResult(task.Result);
                 }
+                s_diagnosticListener.WriteCommandAfter(operationId, this, _transaction);
             }
         }
 
@@ -2756,8 +2781,11 @@ namespace Microsoft.Data.SqlClient
         {
             SqlClientEventSource.Log.TryCorrelationTraceEvent("SqlCommand.InternalExecuteReaderAsync | API | Correlation | Object Id {0}, Behavior {1}, Activity Id {2}, Client Connection Id {3}, Command Text '{4}'", ObjectID, (int)behavior, ActivityCorrelator.Current, Connection?.ClientConnectionId, CommandText);
             SqlClientEventSource.Log.TryTraceEvent("SqlCommand.InternalExecuteReaderAsync | API> {0}, Client Connection Id {1}, Command Text = '{2}'", ObjectID, Connection?.ClientConnectionId, CommandText);
-            SqlConnection.ExecutePermission.Demand();
             Guid operationId = default(Guid);
+            if (!_parentOperationStarted)
+            {
+                operationId = s_diagnosticListener.WriteCommandBefore(this, _transaction);
+            }
 
             // connection can be used as state in RegisterForConnectionCloseNotification continuation
             // to avoid an allocation so use it as the state value if possible but it can be changed if
@@ -2819,6 +2847,11 @@ namespace Microsoft.Data.SqlClient
             }
             catch (Exception e)
             {
+                if (!_parentOperationStarted)
+                {
+                    s_diagnosticListener.WriteCommandError(operationId, this, _transaction, e);
+                }
+
                 source.SetException(e);
                 context?.Dispose();
             }
@@ -2857,6 +2890,11 @@ namespace Microsoft.Data.SqlClient
 
         private Task<object> InternalExecuteScalarAsync(CancellationToken cancellationToken)
         {
+            SqlClientEventSource.Log.TryCorrelationTraceEvent("SqlCommand.InternalExecuteScalarAsync | API | Correlation | Object Id {0}, Activity Id {1}, Client Connection Id {2}, Command Text '{3}'", ObjectID, ActivityCorrelator.Current, Connection?.ClientConnectionId, CommandText);
+            SqlClientEventSource.Log.TryTraceEvent("SqlCommand.InternalExecuteScalarAsync | API> {0}, Client Connection Id {1}, Command Text = '{2}'", ObjectID, Connection?.ClientConnectionId, CommandText);
+            _parentOperationStarted = true;
+            Guid operationId = s_diagnosticListener.WriteCommandBefore(this, _transaction);
+
             return ExecuteReaderAsync(cancellationToken).ContinueWith((executeTask) =>
             {
                 TaskCompletionSource<object> source = new TaskCompletionSource<object>();
@@ -2866,6 +2904,7 @@ namespace Microsoft.Data.SqlClient
                 }
                 else if (executeTask.IsFaulted)
                 {
+                    s_diagnosticListener.WriteCommandError(operationId, this, _transaction, executeTask.Exception.InnerException);
                     source.SetException(executeTask.Exception.InnerException);
                 }
                 else
@@ -2884,6 +2923,7 @@ namespace Microsoft.Data.SqlClient
                                 else if (readTask.IsFaulted)
                                 {
                                     reader.Dispose();
+                                    s_diagnosticListener.WriteCommandError(operationId, this, _transaction, readTask.Exception.InnerException);
                                     source.SetException(readTask.Exception.InnerException);
                                 }
                                 else
@@ -2911,10 +2951,12 @@ namespace Microsoft.Data.SqlClient
                                     }
                                     if (exception != null)
                                     {
+                                        s_diagnosticListener.WriteCommandError(operationId, this, _transaction, exception);
                                         source.SetException(exception);
                                     }
                                     else
                                     {
+                                        s_diagnosticListener.WriteCommandAfter(operationId, this, _transaction);
                                         source.SetResult(result);
                                     }
                                 }
@@ -2928,8 +2970,97 @@ namespace Microsoft.Data.SqlClient
                         TaskScheduler.Default
                     );
                 }
+                _parentOperationStarted = false;
                 return source.Task;
             }, TaskScheduler.Default).Unwrap();
+        }
+
+        internal Task<object> ExecuteScalarBatchAsync(CancellationToken cancellationToken)
+        {
+            _parentOperationStarted = true;
+            Guid operationId = s_diagnosticListener.WriteCommandBefore(this, _transaction);
+
+            return ExecuteReaderAsync(cancellationToken).ContinueWith((executeTask) =>
+            {
+                TaskCompletionSource<object> source = new TaskCompletionSource<object>();
+                if (executeTask.IsCanceled)
+                {
+                    source.SetCanceled();
+                }
+                else if (executeTask.IsFaulted)
+                {
+                    s_diagnosticListener.WriteCommandError(operationId, this, _transaction, executeTask.Exception.InnerException);
+                    source.SetException(executeTask.Exception.InnerException);
+                }
+                else
+                {
+                    SqlDataReader reader = executeTask.Result;
+                    ExecuteScalarUntilEndAsync(reader, cancellationToken).ContinueWith(
+                        (readTask) =>
+                        {
+                            try
+                            {
+                                if (readTask.IsCanceled)
+                                {
+                                    reader.Dispose();
+                                    source.SetCanceled();
+                                }
+                                else if (readTask.IsFaulted)
+                                {
+                                    reader.Dispose();
+                                    s_diagnosticListener.WriteCommandError(operationId, this, _transaction, readTask.Exception.InnerException);
+                                    source.SetException(readTask.Exception.InnerException);
+                                }
+                                else
+                                {
+                                    Exception exception = null;
+                                    object result = null;
+                                    try
+                                    {
+                                        result = readTask.Result;
+                                    }
+                                    finally
+                                    {
+                                        reader.Dispose();
+                                    }
+                                    if (exception != null)
+                                    {
+                                        s_diagnosticListener.WriteCommandError(operationId, this, _transaction, exception);
+                                        source.SetException(exception);
+                                    }
+                                    else
+                                    {
+                                        s_diagnosticListener.WriteCommandAfter(operationId, this, _transaction);
+                                        source.SetResult(result);
+                                    }
+                                }
+                            }
+                            catch (Exception e)
+                            {
+                                // exception thrown by Dispose...
+                                source.SetException(e);
+                            }
+                        },
+                        TaskScheduler.Default
+                    );
+                }
+                _parentOperationStarted = false;
+                return source.Task;
+            }, TaskScheduler.Default).Unwrap();
+        }
+
+        private async Task<object> ExecuteScalarUntilEndAsync(SqlDataReader reader, CancellationToken cancellationToken)
+        {
+            object retval = null;
+            do
+            {
+                if (await reader.ReadAsync(cancellationToken).ConfigureAwait(false) && reader.FieldCount > 0)
+                {
+                    retval = reader.GetValue(0); // no async untyped value getter, this will work ok as long as the value is in the current packet
+                }
+            }
+            while (_batchRPCMode && !cancellationToken.IsCancellationRequested && await reader.NextResultAsync(cancellationToken).ConfigureAwait(false));
+            return retval;
         }
 
         /// <include file='../../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/ExecuteXmlReaderAsync[@name="default"]/*'/>
@@ -2951,8 +3082,7 @@ namespace Microsoft.Data.SqlClient
         private Task<XmlReader> InternalExecuteXmlReaderAsync(CancellationToken cancellationToken)
         {
             SqlClientEventSource.Log.TryCorrelationTraceEvent("SqlCommand.InternalExecuteXmlReaderAsync | API | Correlation | Object Id {0}, Activity Id {1}, Client Connection Id {2}, Command Text '{3}'", ObjectID, ActivityCorrelator.Current, Connection?.ClientConnectionId, CommandText);
-            SqlConnection.ExecutePermission.Demand();
-            Guid operationId = Guid.Empty;
+            Guid operationId = s_diagnosticListener.WriteCommandBefore(this, _transaction);
 
             // connection can be used as state in RegisterForConnectionCloseNotification continuation
             // to avoid an allocation so use it as the state value if possible but it can be changed if
@@ -2987,11 +3117,11 @@ namespace Microsoft.Data.SqlClient
                 returnedTask = RegisterForConnectionCloseNotification(returnedTask);
 
                 Task<XmlReader>.Factory.FromAsync(
-                    beginMethod: static (AsyncCallback callback, object stateObject) =>
+                    beginMethod: static (AsyncCallback callback, object stateObject) => // with c# 10/NET6 add [StackTraceHidden] to this
                     {
                         return ((ExecuteXmlReaderAsyncCallContext)stateObject).Command.BeginExecuteXmlReaderAsync(callback, stateObject);
                     },
-                    endMethod: static (IAsyncResult asyncResult) =>
+                    endMethod: static (IAsyncResult asyncResult) => // with c# 10/NET6 add [StackTraceHidden] to this
                     {
                         return ((ExecuteXmlReaderAsyncCallContext)asyncResult.AsyncState).Command.EndExecuteXmlReaderAsync(asyncResult);
                     },
@@ -3012,6 +3142,7 @@ namespace Microsoft.Data.SqlClient
             }
             catch (Exception e)
             {
+                s_diagnosticListener.WriteCommandError(operationId, this, _transaction, e);
                 source.SetException(e);
             }
 
@@ -3023,6 +3154,7 @@ namespace Microsoft.Data.SqlClient
             if (task.IsFaulted)
             {
                 Exception e = task.Exception.InnerException;
+                s_diagnosticListener.WriteCommandError(operationId, this, _transaction, e);
                 source.SetException(e);
             }
             else
@@ -3035,6 +3167,7 @@ namespace Microsoft.Data.SqlClient
                 {
                     source.SetResult(task.Result);
                 }
+                s_diagnosticListener.WriteCommandAfter(operationId, this, _transaction);
             }
         }
 
@@ -3517,23 +3650,7 @@ namespace Microsoft.Data.SqlClient
         // Check to see if notifications auto enlistment is turned on. Enlist if so.
         private void CheckNotificationStateAndAutoEnlist()
         {
-            // First, if auto-enlist is on, check server version and then obtain context if
-            // present.  If so, auto enlist to the dependency ID given in the context data.
-            if (NotificationAutoEnlist)
-            {
-                string notifyContext = SqlNotificationContext();
-                if (!string.IsNullOrEmpty(notifyContext))
-                {
-                    // Map to dependency by ID set in context data.
-                    SqlDependency dependency = SqlDependencyPerAppDomainDispatcher.SingletonInstance.LookupDependencyEntry(notifyContext);
-
-                    if (dependency != null)
-                    {
-                        // Add this command to the dependency.
-                        dependency.AddCommandDependency(this);
-                    }
-                }
-            }
+            // Auto-enlist not supported in Core
 
             // If we have a notification with a dependency, setup the notification options at this time.
 
@@ -3583,16 +3700,6 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-        [System.Security.Permissions.SecurityPermission(SecurityAction.Assert, Infrastructure = true)]
-        static internal string SqlNotificationContext()
-        {
-            SqlConnection.VerifyExecutePermission();
-
-            // since this information is protected, follow it so that it is not exposed to the user.
-            // SQLBU 329633, SQLBU 329637
-            return (System.Runtime.Remoting.Messaging.CallContext.GetData("MS.SqlDependencyCookie") as string);
-        }
-
         // Tds-specific logic for ExecuteNonQuery run handling
         private Task RunExecuteNonQueryTds(string methodName, bool isAsync, int timeout, bool asyncWrite)
         {
@@ -3610,29 +3717,7 @@ namespace Microsoft.Data.SqlClient
                         TaskCompletionSource<object> completion = new TaskCompletionSource<object>();
                         _activeConnection.RegisterWaitingForReconnect(completion.Task);
                         _reconnectionCompletionSource = completion;
-                        CancellationTokenSource timeoutCTS = new CancellationTokenSource();
-                        AsyncHelper.SetTimeoutException(completion, timeout, static () => SQL.CR_ReconnectTimeout(), timeoutCTS.Token);
-                        AsyncHelper.ContinueTask(reconnectTask, completion,
-                            () =>
-                            {
-                                if (completion.Task.IsCompleted)
-                                {
-                                    return;
-                                }
-                                Interlocked.CompareExchange(ref _reconnectionCompletionSource, null, completion);
-                                timeoutCTS.Cancel();
-                                Task subTask = RunExecuteNonQueryTds(methodName, isAsync, TdsParserStaticMethods.GetRemainingTimeout(timeout, reconnectionStart), asyncWrite);
-                                if (subTask == null)
-                                {
-                                    completion.SetResult(null);
-                                }
-                                else
-                                {
-                                    AsyncHelper.ContinueTaskWithState(subTask, completion, completion, static (object state) => ((TaskCompletionSource<object>)state).SetResult(null));
-                                }
-                            },
-                            connectionToAbort: _activeConnection
-                        );
+                        RunExecuteNonQueryTdsSetupReconnnectContinuation(methodName, isAsync, timeout, asyncWrite, reconnectTask, reconnectionStart, completion);
                         return completion.Task;
                     }
                     else
@@ -3689,6 +3774,36 @@ namespace Microsoft.Data.SqlClient
                 }
             }
             return null;
+        }
+
+        // This is in its own method to avoid always allocating the lambda in RunExecuteNonQueryTds, cannot use ContinueTaskWithState because of MarshalByRef and the CompareExchange
+        private void RunExecuteNonQueryTdsSetupReconnnectContinuation(string methodName, bool isAsync, int timeout, bool asyncWrite, Task reconnectTask, long reconnectionStart, TaskCompletionSource<object> completion)
+        {
+            CancellationTokenSource timeoutCTS = new CancellationTokenSource();
+            AsyncHelper.SetTimeoutException(completion, timeout, static () => SQL.CR_ReconnectTimeout(), timeoutCTS.Token);
+            AsyncHelper.ContinueTask(reconnectTask, completion,
+                () =>
+                {
+                    if (completion.Task.IsCompleted)
+                    {
+                        return;
+                    }
+                    Interlocked.CompareExchange(ref _reconnectionCompletionSource, null, completion);
+                    timeoutCTS.Cancel();
+                    Task subTask = RunExecuteNonQueryTds(methodName, isAsync, TdsParserStaticMethods.GetRemainingTimeout(timeout, reconnectionStart), asyncWrite);
+                    if (subTask == null)
+                    {
+                        completion.SetResult(null);
+                    }
+                    else
+                    {
+                        AsyncHelper.ContinueTaskWithState(subTask, completion,
+                            state: completion,
+                            onSuccess: static (object state) => ((TaskCompletionSource<object>)state).SetResult(null)
+                        );
+                    }
+                }
+            );
         }
 
         /// <summary>
@@ -3823,6 +3938,7 @@ namespace Microsoft.Data.SqlClient
             // Used in _batchRPCMode to maintain a map of describe parameter encryption RPC requests (Keys) and their corresponding original RPC requests (Values).
             ReadOnlyDictionary<_SqlRPC, _SqlRPC> describeParameterEncryptionRpcOriginalRpcMap = null;
 
+
             try
             {
                 try
@@ -3869,69 +3985,13 @@ namespace Microsoft.Data.SqlClient
                         // Mark that we should not process the finally block since we have async execution pending.
                         // Note that this should be done outside the task's continuation delegate.
                         processFinallyBlock = false;
-                        returnTask = AsyncHelper.CreateContinuationTask(fetchInputParameterEncryptionInfoTask, () =>
-                        {
-                            bool processFinallyBlockAsync = true;
-                            bool decrementAsyncCountInFinallyBlockAsync = true;
-
-                            try
-                            {
-                                // Check for any exceptions on network write, before reading.
-                                CheckThrowSNIException();
-
-                                // If it is async, then TryFetchInputParameterEncryptionInfo-> RunExecuteReaderTds would have incremented the async count.
-                                // Decrement it when we are about to complete async execute reader.
-                                SqlInternalConnectionTds internalConnectionTds = _activeConnection.GetOpenTdsConnection();
-                                if (internalConnectionTds != null)
-                                {
-                                    internalConnectionTds.DecrementAsyncCount();
-                                    decrementAsyncCountInFinallyBlockAsync = false;
-                                }
-
-                                // Complete executereader.
-                                describeParameterEncryptionDataReader = CompleteAsyncExecuteReader(forDescribeParameterEncryption: true);
-                                Debug.Assert(_stateObj == null, "non-null state object in PrepareForTransparentEncryption.");
-
-                                // Read the results of describe parameter encryption.
-                                ReadDescribeEncryptionParameterResults(
-                                    describeParameterEncryptionDataReader,
-                                    describeParameterEncryptionRpcOriginalRpcMap,
-                                    isRetry);
-
-#if DEBUG
-                                // Failpoint to force the thread to halt to simulate cancellation of SqlCommand.
-                                if (_sleepAfterReadDescribeEncryptionParameterResults)
-                                {
-                                    Thread.Sleep(10000);
-                                }
-#endif //DEBUG
-                            }
-                            catch (Exception e)
-                            {
-                                processFinallyBlockAsync = ADP.IsCatchableExceptionType(e);
-                                throw;
-                            }
-                            finally
-                            {
-                                PrepareTransparentEncryptionFinallyBlock(closeDataReader: processFinallyBlockAsync,
-                                                                            decrementAsyncCount: decrementAsyncCountInFinallyBlockAsync,
-                                                                            clearDataStructures: processFinallyBlockAsync,
-                                                                            wasDescribeParameterEncryptionNeeded: describeParameterEncryptionNeeded,
-                                                                            describeParameterEncryptionRpcOriginalRpcMap: describeParameterEncryptionRpcOriginalRpcMap,
-                                                                            describeParameterEncryptionDataReader: describeParameterEncryptionDataReader);
-                            }
-                        },
-                        onFailure: ((exception) =>
-                        {
-                            if (CachedAsyncState != null)
-                            {
-                                CachedAsyncState.ResetAsyncState();
-                            }
-                            if (exception != null)
-                            {
-                                throw exception;
-                            }
-                        }));
+                        describeParameterEncryptionDataReader = GetParameterEncryptionDataReader(
+                            out returnTask,
+                            fetchInputParameterEncryptionInfoTask,
+                            describeParameterEncryptionDataReader,
+                            describeParameterEncryptionRpcOriginalRpcMap,
+                            describeParameterEncryptionNeeded,
+                            isRetry);
 
                         decrementAsyncCountInFinallyBlock = false;
                     }
@@ -3943,62 +4003,22 @@ namespace Microsoft.Data.SqlClient
                             // Mark that we should not process the finally block since we have async execution pending.
                             // Note that this should be done outside the task's continuation delegate.
                             processFinallyBlock = false;
-                            returnTask = Task.Run(() =>
-                            {
-                                bool processFinallyBlockAsync = true;
-                                bool decrementAsyncCountInFinallyBlockAsync = true;
-
-                                try
-                                {
-
-                                    // Check for any exceptions on network write, before reading.
-                                    CheckThrowSNIException();
-
-                                    // If it is async, then TryFetchInputParameterEncryptionInfo-> RunExecuteReaderTds would have incremented the async count.
-                                    // Decrement it when we are about to complete async execute reader.
-                                    SqlInternalConnectionTds internalConnectionTds = _activeConnection.GetOpenTdsConnection();
-                                    if (internalConnectionTds != null)
-                                    {
-                                        internalConnectionTds.DecrementAsyncCount();
-                                        decrementAsyncCountInFinallyBlockAsync = false;
-                                    }
-
-                                    // Complete executereader.
-                                    describeParameterEncryptionDataReader = CompleteAsyncExecuteReader(forDescribeParameterEncryption: true);
-                                    Debug.Assert(_stateObj == null, "non-null state object in PrepareForTransparentEncryption.");
-
-                                    // Read the results of describe parameter encryption.
-                                    ReadDescribeEncryptionParameterResults(describeParameterEncryptionDataReader, describeParameterEncryptionRpcOriginalRpcMap, isRetry);
-#if DEBUG
-                                    // Failpoint to force the thread to halt to simulate cancellation of SqlCommand.
-                                    if (_sleepAfterReadDescribeEncryptionParameterResults)
-                                    {
-                                        Thread.Sleep(10000);
-                                    }
-#endif
-                                }
-                                catch (Exception e)
-                                {
-                                    processFinallyBlockAsync = ADP.IsCatchableExceptionType(e);
-                                    throw;
-                                }
-                                finally
-                                {
-                                    PrepareTransparentEncryptionFinallyBlock(closeDataReader: processFinallyBlockAsync,
-                                                                                decrementAsyncCount: decrementAsyncCountInFinallyBlockAsync,
-                                                                                clearDataStructures: processFinallyBlockAsync,
-                                                                                wasDescribeParameterEncryptionNeeded: describeParameterEncryptionNeeded,
-                                                                                describeParameterEncryptionRpcOriginalRpcMap: describeParameterEncryptionRpcOriginalRpcMap,
-                                                                                describeParameterEncryptionDataReader: describeParameterEncryptionDataReader);
-                                }
-                            });
+                            describeParameterEncryptionDataReader = GetParameterEncryptionDataReaderAsync(
+                                out returnTask,
+                                describeParameterEncryptionDataReader,
+                                describeParameterEncryptionRpcOriginalRpcMap,
+                                describeParameterEncryptionNeeded,
+                                isRetry);
 
                             decrementAsyncCountInFinallyBlock = false;
                         }
                         else
                         {
                             // For synchronous execution, read the results of describe parameter encryption here.
-                            ReadDescribeEncryptionParameterResults(describeParameterEncryptionDataReader, describeParameterEncryptionRpcOriginalRpcMap, isRetry);
+                            ReadDescribeEncryptionParameterResults(
+                                describeParameterEncryptionDataReader,
+                                describeParameterEncryptionRpcOriginalRpcMap,
+                                isRetry);
                         }
 
 #if DEBUG
@@ -4027,7 +4047,6 @@ namespace Microsoft.Data.SqlClient
                                            describeParameterEncryptionDataReader: describeParameterEncryptionDataReader);
                 }
             }
-            // @TODO: CER Exception Handling was removed here (see GH#3581)
             catch (Exception e)
             {
                 if (CachedAsyncState != null)
@@ -4042,6 +4061,135 @@ namespace Microsoft.Data.SqlClient
 
                 throw;
             }
+        }
+
+        private SqlDataReader GetParameterEncryptionDataReader(out Task returnTask, Task fetchInputParameterEncryptionInfoTask,
+            SqlDataReader describeParameterEncryptionDataReader,
+            ReadOnlyDictionary<_SqlRPC, _SqlRPC> describeParameterEncryptionRpcOriginalRpcMap, bool describeParameterEncryptionNeeded, bool isRetry)
+        {
+            returnTask = AsyncHelper.CreateContinuationTaskWithState(fetchInputParameterEncryptionInfoTask, this,
+                (object state) =>
+                {
+                    SqlCommand command = (SqlCommand)state;
+                    bool processFinallyBlockAsync = true;
+                    bool decrementAsyncCountInFinallyBlockAsync = true;
+
+                    try
+                    {
+                        // Check for any exceptions on network write, before reading.
+                        command.CheckThrowSNIException();
+
+                        // If it is async, then TryFetchInputParameterEncryptionInfo-> RunExecuteReaderTds would have incremented the async count.
+                        // Decrement it when we are about to complete async execute reader.
+                        SqlInternalConnectionTds internalConnectionTds = command._activeConnection.GetOpenTdsConnection();
+                        if (internalConnectionTds != null)
+                        {
+                            internalConnectionTds.DecrementAsyncCount();
+                            decrementAsyncCountInFinallyBlockAsync = false;
+                        }
+
+                        // Complete executereader.
+                        describeParameterEncryptionDataReader = command.CompleteAsyncExecuteReader(forDescribeParameterEncryption: true);
+                        Debug.Assert(command._stateObj == null, "non-null state object in PrepareForTransparentEncryption.");
+
+                        // Read the results of describe parameter encryption.
+                        command.ReadDescribeEncryptionParameterResults(describeParameterEncryptionDataReader, describeParameterEncryptionRpcOriginalRpcMap, isRetry);
+
+#if DEBUG
+                        // Failpoint to force the thread to halt to simulate cancellation of SqlCommand.
+                        if (_sleepAfterReadDescribeEncryptionParameterResults)
+                        {
+                            Thread.Sleep(10000);
+                        }
+#endif
+                    }
+                    catch (Exception e)
+                    {
+                        processFinallyBlockAsync = ADP.IsCatchableExceptionType(e);
+                        throw;
+                    }
+                    finally
+                    {
+                        command.PrepareTransparentEncryptionFinallyBlock(closeDataReader: processFinallyBlockAsync,
+                            decrementAsyncCount: decrementAsyncCountInFinallyBlockAsync,
+                            clearDataStructures: processFinallyBlockAsync,
+                            wasDescribeParameterEncryptionNeeded: describeParameterEncryptionNeeded,
+                            describeParameterEncryptionRpcOriginalRpcMap: describeParameterEncryptionRpcOriginalRpcMap,
+                            describeParameterEncryptionDataReader: describeParameterEncryptionDataReader);
+                    }
+                },
+                onFailure: static (Exception exception, object state) =>
+                {
+                    SqlCommand command = (SqlCommand)state;
+                    if (command.CachedAsyncState != null)
+                    {
+                        command.CachedAsyncState.ResetAsyncState();
+                    }
+
+                    if (exception != null)
+                    {
+                        throw exception;
+                    }
+                }
+            );
+
+            return describeParameterEncryptionDataReader;
+        }
+
+        private SqlDataReader GetParameterEncryptionDataReaderAsync(out Task returnTask,
+            SqlDataReader describeParameterEncryptionDataReader,
+            ReadOnlyDictionary<_SqlRPC, _SqlRPC> describeParameterEncryptionRpcOriginalRpcMap, bool describeParameterEncryptionNeeded, bool isRetry)
+        {
+            returnTask = Task.Run(() =>
+            {
+                bool processFinallyBlockAsync = true;
+                bool decrementAsyncCountInFinallyBlockAsync = true;
+                
+                try
+                {
+                    // Check for any exceptions on network write, before reading.
+                    CheckThrowSNIException();
+
+                    // If it is async, then TryFetchInputParameterEncryptionInfo-> RunExecuteReaderTds would have incremented the async count.
+                    // Decrement it when we are about to complete async execute reader.
+                    SqlInternalConnectionTds internalConnectionTds = _activeConnection.GetOpenTdsConnection();
+                    if (internalConnectionTds != null)
+                    {
+                        internalConnectionTds.DecrementAsyncCount();
+                        decrementAsyncCountInFinallyBlockAsync = false;
+                    }
+
+                    // Complete executereader.
+                    describeParameterEncryptionDataReader = CompleteAsyncExecuteReader(forDescribeParameterEncryption: true);
+                    Debug.Assert(_stateObj == null, "non-null state object in PrepareForTransparentEncryption.");
+
+                    // Read the results of describe parameter encryption.
+                    ReadDescribeEncryptionParameterResults(describeParameterEncryptionDataReader,
+                        describeParameterEncryptionRpcOriginalRpcMap, isRetry);
+#if DEBUG
+                    // Failpoint to force the thread to halt to simulate cancellation of SqlCommand.
+                    if (_sleepAfterReadDescribeEncryptionParameterResults)
+                    {
+                        Thread.Sleep(10000);
+                    }
+#endif
+                }
+                catch (Exception e)
+                {
+                    processFinallyBlockAsync = ADP.IsCatchableExceptionType(e);
+                    throw;
+                }
+                finally
+                {
+                    PrepareTransparentEncryptionFinallyBlock(closeDataReader: processFinallyBlockAsync,
+                        decrementAsyncCount: decrementAsyncCountInFinallyBlockAsync,
+                        clearDataStructures: processFinallyBlockAsync,
+                        wasDescribeParameterEncryptionNeeded: describeParameterEncryptionNeeded,
+                        describeParameterEncryptionRpcOriginalRpcMap: describeParameterEncryptionRpcOriginalRpcMap,
+                        describeParameterEncryptionDataReader: describeParameterEncryptionDataReader);
+                }
+            });
+            return describeParameterEncryptionDataReader;
         }
 
         /// <summary>
@@ -4718,7 +4866,6 @@ namespace Microsoft.Data.SqlClient
 
             CheckNotificationStateAndAutoEnlist(); // Only call after validate - requires non null connection!
 
-            // This section needs to occur AFTER ValidateCommand - otherwise it will AV without a connection.
             SqlStatistics statistics = Statistics;
             if (statistics != null)
             {
@@ -4838,7 +4985,6 @@ namespace Microsoft.Data.SqlClient
             {
                 return RunExecuteReaderTds(cmdBehavior, runBehavior, returnStream, isAsync, timeout, out task, asyncWrite && isAsync, isRetry: isRetry);
             }
-            // @TODO: CER Exception Handling was removed here (see GH#3581)
         }
 
 
@@ -4878,7 +5024,7 @@ namespace Microsoft.Data.SqlClient
                         }
                         else
                         {
-                            AsyncHelper.ContinueTaskWithState(subTask, completion, completion, static (object state2) => ((TaskCompletionSource<object>)state2).SetResult(null));
+                            AsyncHelper.ContinueTaskWithState(subTask, completion, completion, static (object state) => ((TaskCompletionSource<object>)state).SetResult(null));
                         }
                     },
                     onFailure: static (Exception exception, object state) =>
@@ -4889,9 +5035,11 @@ namespace Microsoft.Data.SqlClient
                             throw exception;
                         }
                     },
-                    onCancellation: static (object state) => ((SqlCommand)state).CachedAsyncState?.ResetAsyncState(),
-                    connectionToDoom: null,
-                    connectionToAbort: _activeConnection);
+                    onCancellation: static (object state) =>
+                    {
+                        ((SqlCommand)state).CachedAsyncState?.ResetAsyncState();
+                    }
+                );
                 task = completion.Task;
                 return ds;
             }
@@ -4977,30 +5125,7 @@ namespace Microsoft.Data.SqlClient
                     TaskCompletionSource<object> completion = new TaskCompletionSource<object>();
                     _activeConnection.RegisterWaitingForReconnect(completion.Task);
                     _reconnectionCompletionSource = completion;
-                    CancellationTokenSource timeoutCTS = new CancellationTokenSource();
-                    AsyncHelper.SetTimeoutException(completion, timeout, static () => SQL.CR_ReconnectTimeout(), timeoutCTS.Token);
-                    AsyncHelper.ContinueTask(reconnectTask, completion,
-                        () =>
-                        {
-                            if (completion.Task.IsCompleted)
-                            {
-                                return;
-                            }
-                            Interlocked.CompareExchange(ref _reconnectionCompletionSource, null, completion);
-                            timeoutCTS.Cancel();
-                            Task subTask;
-                            RunExecuteReaderTds(cmdBehavior, runBehavior, returnStream, isAsync, TdsParserStaticMethods.GetRemainingTimeout(timeout, reconnectionStart), out subTask, asyncWrite, isRetry, ds);
-                            if (subTask == null)
-                            {
-                                completion.SetResult(null);
-                            }
-                            else
-                            {
-                                AsyncHelper.ContinueTaskWithState(subTask, completion, completion, static (object state) => ((TaskCompletionSource<object>)state).SetResult(null));
-                            }
-                        },
-                        connectionToAbort: _activeConnection
-                    );
+                    RunExecuteReaderTdsSetupReconnectContinuation(cmdBehavior, runBehavior, returnStream, isAsync, timeout, asyncWrite, isRetry, ds, reconnectTask, reconnectionStart, completion);
                     task = completion.Task;
                     return ds;
                 }
@@ -5129,14 +5254,12 @@ namespace Microsoft.Data.SqlClient
                         BuildExecuteSql(cmdBehavior, null, _parameters, ref rpc);
                     }
 
-                    // if 2000, then set NOMETADATA_UNLESSCHANGED flag
                     rpc.options = TdsEnums.RPC_NOMETADATA;
                     if (returnStream)
                     {
                         SqlClientEventSource.Log.TryTraceEvent("SqlCommand.RunExecuteReaderTds | Info | Object Id {0}, Activity Id {1}, Client Connection Id {2}, Command executed as RPC, RPC Name '{3}' ", ObjectID, ActivityCorrelator.Current, Connection?.ClientConnectionId, rpc?.rpcName);
                     }
 
-                    // TODO: Medusa: Unprepare only happens for SQL 7.0 which may be broken anyway (it's not re-prepared). Consider removing the reset here if we're really dropping 7.0 support.
                     Debug.Assert(_rpcArrayOf1[0] == rpc);
                     writeTask = _stateObj.Parser.TdsExecuteRPC(this, _rpcArrayOf1, timeout, inSchema, this.Notification, _stateObj, CommandType.StoredProcedure == CommandType, sync: !asyncWrite);
                 }
@@ -5161,7 +5284,7 @@ namespace Microsoft.Data.SqlClient
                         Task executeTask = _stateObj.Parser.TdsExecuteSQLBatch(optionSettings, timeout, this.Notification, _stateObj, sync: true);
                         Debug.Assert(executeTask == null, "Shouldn't get a task when doing sync writes");
                         Debug.Assert(_stateObj._syncOverAsync, "Should not attempt pends in a synchronous call");
-                        TdsOperationStatus result = _stateObj.Parser.TryRun(RunBehavior.UntilDone, this, null, null, _stateObj, out _);
+                        TdsOperationStatus result = _stateObj.Parser.TryRun(RunBehavior.UntilDone, this, null, null, _stateObj, out bool dataReady);
                         if (result != TdsOperationStatus.Done)
                         {
                             throw SQL.SynchronousCallMayNotPend();
@@ -5182,15 +5305,7 @@ namespace Microsoft.Data.SqlClient
                     decrementAsyncCountOnFailure = false;
                     if (writeTask != null)
                     {
-                        task = AsyncHelper.CreateContinuationTask(writeTask, () =>
-                        {
-                            _activeConnection.GetOpenTdsConnection(); // it will throw if connection is closed
-                            CachedAsyncState.SetAsyncReaderState(ds, runBehavior, optionSettings);
-                        },
-                                 onFailure: (exc) =>
-                                 {
-                                     _activeConnection.GetOpenTdsConnection().DecrementAsyncCount();
-                                 });
+                        task = RunExecuteReaderTdsSetupContinuation(runBehavior, ds, optionSettings, writeTask);
                     }
                     else
                     {
@@ -5229,6 +5344,57 @@ namespace Microsoft.Data.SqlClient
             Debug.Assert(isAsync || _stateObj == null, "non-null state object in RunExecuteReader");
             return ds;
         }
+
+        private Task RunExecuteReaderTdsSetupContinuation(RunBehavior runBehavior, SqlDataReader ds, string optionSettings, Task writeTask)
+        {
+            Task task = AsyncHelper.CreateContinuationTaskWithState(
+                task: writeTask,
+                state: _activeConnection,
+                onSuccess: (object state) =>
+                {
+                    SqlConnection sqlConnection = (SqlConnection)state;
+                    sqlConnection.GetOpenTdsConnection(); // it will throw if connection is closed
+                    CachedAsyncState.SetAsyncReaderState(ds, runBehavior, optionSettings);
+                },
+                onFailure: static (Exception exc, object state) =>
+                {
+                    ((SqlConnection)state).GetOpenTdsConnection().DecrementAsyncCount();
+                }
+            );
+            return task;
+        }
+
+        // This is in its own method to avoid always allocating the lambda in RunExecuteReaderTds
+        private void RunExecuteReaderTdsSetupReconnectContinuation(CommandBehavior cmdBehavior, RunBehavior runBehavior, bool returnStream, bool isAsync, int timeout, bool asyncWrite, bool isRetry, SqlDataReader ds, Task reconnectTask, long reconnectionStart, TaskCompletionSource<object> completion)
+        {
+            CancellationTokenSource timeoutCTS = new CancellationTokenSource();
+            AsyncHelper.SetTimeoutException(completion, timeout, static () => SQL.CR_ReconnectTimeout(), timeoutCTS.Token);
+            AsyncHelper.ContinueTask(reconnectTask, completion,
+                () =>
+                {
+                    if (completion.Task.IsCompleted)
+                    {
+                        return;
+                    }
+                    Interlocked.CompareExchange(ref _reconnectionCompletionSource, null, completion);
+                    timeoutCTS.Cancel();
+                    Task subTask;
+                    RunExecuteReaderTds(cmdBehavior, runBehavior, returnStream, isAsync, TdsParserStaticMethods.GetRemainingTimeout(timeout, reconnectionStart), out subTask, asyncWrite, isRetry, ds);
+                    if (subTask == null)
+                    {
+                        completion.SetResult(null);
+                    }
+                    else
+                    {
+                        AsyncHelper.ContinueTaskWithState(subTask, completion,
+                             state: completion,
+                             onSuccess: static (object state) => ((TaskCompletionSource<object>)state).SetResult(null)
+                         );
+                    }
+                }
+            );
+        }
+
 
         private SqlDataReader CompleteAsyncExecuteReader(bool isInternal = false, bool forDescribeParameterEncryption = false)
         {
@@ -5302,7 +5468,14 @@ namespace Microsoft.Data.SqlClient
 
                         if (ds != null)
                         {
-                            ds.Close();
+                            try
+                            {
+                                ds.Close();
+                            }
+                            catch (Exception exClose)
+                            {
+                                Debug.WriteLine("Received this exception from SqlDataReader.Close() while in another catch block: " + exClose.ToString());
+                            }
                         }
                     }
                     throw;
@@ -5355,7 +5528,14 @@ namespace Microsoft.Data.SqlClient
                             _execType = EXECTYPE.PREPAREPENDING; // reset execution type to pending
                         }
 
-                        ds.Close();
+                        try
+                        {
+                            ds.Close();
+                        }
+                        catch (Exception exClose)
+                        {
+                            Debug.WriteLine("Received this exception from SqlDataReader.Close() while in another catch block: " + exClose.ToString());
+                        }
                     }
 
                     throw;
@@ -5374,7 +5554,7 @@ namespace Microsoft.Data.SqlClient
         object ICloneable.Clone() =>
             Clone();
 
-        private Task<T> RegisterForConnectionCloseNotification<T>(Task<T> outterTask)
+        private Task<T> RegisterForConnectionCloseNotification<T>(Task<T> outerTask)
         {
             SqlConnection connection = _activeConnection;
             if (connection == null)
@@ -5383,7 +5563,7 @@ namespace Microsoft.Data.SqlClient
                 throw ADP.ClosedConnectionError();
             }
 
-            return connection.RegisterForConnectionCloseNotification(outterTask, this, SqlReferenceCollection.CommandTag);
+            return connection.RegisterForConnectionCloseNotification(outerTask, this, SqlReferenceCollection.CommandTag);
         }
 
         // validates that a command has commandText and a non-busy open connection
@@ -5434,8 +5614,6 @@ namespace Microsoft.Data.SqlClient
             // close any non MARS dead readers, if applicable, and then throw if still busy.
             // Throw if we have a live reader on this command
             _activeConnection.ValidateConnectionForExecute(method, this);
-            // @TODO: CER Exception Handling was removed here (see GH#3581)
-
             // Check to see if the currently set transaction has completed.  If so,
             // null out our local reference.
             if (_transaction != null && _transaction.Connection == null)
@@ -5466,7 +5644,7 @@ namespace Microsoft.Data.SqlClient
 
         private void ValidateAsyncCommand()
         {
-            if (CachedAsyncState.PendingAsyncOperation)
+            if (CachedAsyncState != null && CachedAsyncState.PendingAsyncOperation)
             {
                 // Enforce only one pending async execute at a time.
                 if (CachedAsyncState.IsActiveConnectionValid(_activeConnection))
@@ -5528,7 +5706,6 @@ namespace Microsoft.Data.SqlClient
         private void ReliablePutStateObject()
         {
             PutStateObject();
-            // @TODO: CER Exception Handling was removed here (see GH#3581)
         }
 
         private void PutStateObject()
@@ -5597,12 +5774,11 @@ namespace Microsoft.Data.SqlClient
 
         internal void OnReturnStatus(int status)
         {
-            if (_inPrepare)
-                return;
-
             // Don't set the return status if this is the status for sp_describe_parameter_encryption.
-            if (IsDescribeParameterEncryptionRPCCurrentlyInProgress)
+            if (_inPrepare || IsDescribeParameterEncryptionRPCCurrentlyInProgress)
+            {
                 return;
+            }
 
             SqlParameterCollection parameters = _parameters;
             if (_batchRPCMode)
@@ -5779,7 +5955,7 @@ namespace Microsoft.Data.SqlClient
                         thisParam.SetSqlBuffer(rec.value);
                     }
 
-                    MetaType mt = MetaType.GetMetaTypeFromSqlDbType(rec.type, rec.IsMultiValued);
+                    MetaType mt = MetaType.GetMetaTypeFromSqlDbType(rec.type, false);
 
                     if (rec.type == SqlDbType.Decimal)
                     {
@@ -5903,17 +6079,6 @@ namespace Microsoft.Data.SqlClient
             rpc.rpcName = null;
             rpc.options = 0;
             rpc.systemParamCount = systemParamCount;
-
-            rpc.recordsAffected = default(int?);
-            rpc.cumulativeRecordsAffected = -1;
-
-            rpc.errorsIndexStart = 0;
-            rpc.errorsIndexEnd = 0;
-            rpc.errors = null;
-
-            rpc.warningsIndexStart = 0;
-            rpc.warningsIndexEnd = 0;
-            rpc.warnings = null;
             rpc.needsFetchParameterEncryptionMetadata = false;
 
             int currentCount = rpc.systemParams?.Length ?? 0;
@@ -5998,7 +6163,7 @@ namespace Microsoft.Data.SqlClient
                                     parts[1] != null // server should not be null or we don't need to remove it
                                 )
                                 {
-                                    parameter.TypeName = QuoteIdentifier(parts, 2, 2);
+                                    parameter.TypeName = QuoteIdentifier(parts.AsSpan(2, 2));
                                 }
                             }
                         }
@@ -6490,10 +6655,10 @@ namespace Microsoft.Data.SqlClient
         private static string ParseAndQuoteIdentifier(string identifier, bool isUdtTypeName)
         {
             string[] strings = SqlParameter.ParseTypeName(identifier, isUdtTypeName);
-            return ADP.BuildMultiPartName(strings);
+            return QuoteIdentifier(strings);
         }
 
-        private static string QuoteIdentifier(string[] strings, int offset, int length)
+        private static string QuoteIdentifier(ReadOnlySpan<string> strings)
         {
             StringBuilder bld = new StringBuilder();
 
@@ -6501,7 +6666,7 @@ namespace Microsoft.Data.SqlClient
             //  with all parts except trimming separators for leading empty names (null or empty strings,
             //  but not whitespace). Separators in the middle should be added, even if the name part is 
             //  null/empty, to maintain proper location of the parts.
-            for (int i = offset; i < (offset + length); i++)
+            for (int i = 0; i < strings.Length; i++)
             {
                 if (0 < bld.Length)
                 {
@@ -6801,17 +6966,18 @@ namespace Microsoft.Data.SqlClient
         internal SqlException GetErrors(int commandIndex)
         {
             SqlException result = null;
-            int length = (_RPCList[commandIndex].errorsIndexEnd - _RPCList[commandIndex].errorsIndexStart);
+            _SqlRPC rpc = _RPCList[commandIndex];
+            int length = (rpc.errorsIndexEnd - rpc.errorsIndexStart);
             if (0 < length)
             {
                 SqlErrorCollection errors = new SqlErrorCollection();
-                for (int i = _RPCList[commandIndex].errorsIndexStart; i < _RPCList[commandIndex].errorsIndexEnd; ++i)
+                for (int i = rpc.errorsIndexStart; i < rpc.errorsIndexEnd; ++i)
                 {
-                    errors.Add(_RPCList[commandIndex].errors[i]);
+                    errors.Add(rpc.errors[i]);
                 }
-                for (int i = _RPCList[commandIndex].warningsIndexStart; i < _RPCList[commandIndex].warningsIndexEnd; ++i)
+                for (int i = rpc.warningsIndexStart; i < rpc.warningsIndexEnd; ++i)
                 {
-                    errors.Add(_RPCList[commandIndex].warnings[i]);
+                    errors.Add(rpc.warnings[i]);
                 }
                 result = SqlException.CreateException(errors, Connection.ServerVersion, Connection.ClientConnectionId, innerException: null, batchCommand: null);
             }
