--- src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/SqlCommand.stub.cs
@@ -1,15 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-// @TODO: This is only a stub class for removing clearing errors while merging other files.
-
-namespace Microsoft.Data.SqlClient
-{
-    public class SqlCommand
-    {
-        public SqlConnection Connection { get; set; }
-
-        internal SqlStatistics Statistics { get; set; }
-    }
-}
